Create TextInput
Created a text input component that can be easily integrated into any
part of the app. The new component accepts several props to enhance
reusability. The primary props include:

value: A prop to pass the initial value of the input field.

onChange: A function prop that will be triggered whenever the
input value changes.

styles: Props for customizing the styles of the text input,
allowing seamless integration into different app contexts.
