+++ renta_facil_colima/App.jsx
@@ -1,0 +1,91 @@
+/**
+ * Sample React Native App
+ * https://github.com/facebook/react-native
+ *
+ * @format
+ */
+
+import React, { useState } from 'react';
+
+import {
+  SafeAreaView,
+  ScrollView,
+  StatusBar,
+  StyleSheet,
+  Text,
+  TextInput,
+  useColorScheme,
+  View,
+} from 'react-native';
+
+import {
+  Colors,
+  DebugInstructions,
+  Header,
+  LearnMoreLinks,
+  ReloadInstructions,
+} from 'react-native/Libraries/NewAppScreen';
+import Textinput from './View/Components/Textinput';
+
+
+
+
+
+function App() {
+  const isDarkMode = useColorScheme() === 'dark';
+
+  const backgroundStyle = {
+    backgroundColor: isDarkMode ? Colors.darker : Colors.lighter,
+  };
+  const [nombre, setNombre] = useState('')
+  const [apellidos, setApellidos] = useState('')
+  console.log(nombre);
+  console.log(apellidos);
+
+  return (
+    <SafeAreaView style={backgroundStyle}>
+      <StatusBar
+        barStyle={isDarkMode ? 'light-content' : 'dark-content'}
+        backgroundColor={backgroundStyle.backgroundColor}
+      />
+     <Textinput 
+      altura={40}
+      padding={10}
+      margen={10}
+      placeholder='First name'
+      value={nombre}
+      onChangeText={setNombre}
+     />
+      <Textinput 
+      altura={40}
+      padding={10}
+      margen={10}
+      placeholder='Second name'
+      value={apellidos}
+      onChangeText={setApellidos}
+     />
+     
+    </SafeAreaView>
+  );
+}
+
+const styles = StyleSheet.create({
+  sectionContainer: {
+    marginTop: 32,
+    paddingHorizontal: 24,
+  },
+  sectionTitle: {
+    fontSize: 24,
+    fontWeight: '600',
+  },
+  sectionDescription: {
+    marginTop: 8,
+    fontSize: 18,
+    fontWeight: '400',
+  },
+  highlight: {
+    fontWeight: '700',
+  },
+});
+
+export default App;
